TDS-519: consolidate status change state map
